Move functionality to lib
